automation/requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
packaging==23.0
python-rpm-spec==0.14
requests==2.28.2
smmap==5.0.0
urllib3==1.26.14
